+++ src/Contracts/PruneableEvent.php
@@ -2,6 +2,7 @@
 
 namespace AaronFrancis\Eventable\Contracts;
 
+use AaronFrancis\Eventable\Prune;
 use AaronFrancis\Eventable\PruneConfig;
 
 interface PruneableEvent
@@ -10,5 +11,5 @@ interface PruneableEvent
      * Get the prune configuration for this event type.
      * Return null to skip pruning for this event type.
      */
-    public function prune(): ?PruneConfig;
+    public function prune(): PruneConfig|Prune|null;
 }