+++ tests/Fixtures/CombinedPruneEvent.php
@@ -3,6 +3,7 @@
 namespace AaronFrancis\Eventable\Tests\Fixtures;
 
 use AaronFrancis\Eventable\Contracts\PruneableEvent;
+use AaronFrancis\Eventable\Prune;
 use AaronFrancis\Eventable\PruneConfig;
 use Illuminate\Support\Carbon;
 
@@ -11,18 +12,13 @@ enum CombinedPruneEvent: int implements PruneableEvent
     case KeepLast3OlderThan7Days = 1;
     case KeepLast5NoVaryOnData = 2;
 
-    public function prune(): ?PruneConfig
+    public function prune(): PruneConfig|Prune|null
     {
         return match ($this) {
-            self::KeepLast3OlderThan7Days => new PruneConfig(
-                before: Carbon::now()->subDays(7),
-                keep: 3,
-                varyOnData: true
-            ),
-            self::KeepLast5NoVaryOnData => new PruneConfig(
-                keep: 5,
-                varyOnData: false
-            ),
+            self::KeepLast3OlderThan7Days => Prune::before(Carbon::now()->subDays(7))
+                ->keep(3)
+                ->varyOnData(),
+            self::KeepLast5NoVaryOnData => Prune::keep(5)->dontVaryOnData(),
         };
     }
 }