+++ tests/Fixtures/PruneableTestEvent.php
@@ -3,6 +3,7 @@
 namespace AaronFrancis\Eventable\Tests\Fixtures;
 
 use AaronFrancis\Eventable\Contracts\PruneableEvent;
+use AaronFrancis\Eventable\Prune;
 use AaronFrancis\Eventable\PruneConfig;
 use Illuminate\Support\Carbon;
 
@@ -14,13 +15,13 @@ enum PruneableTestEvent: int implements PruneableEvent
     case KeepLast3VaryOnData = 4;
     case NoPruneConfig = 5;
 
-    public function prune(): ?PruneConfig
+    public function prune(): PruneConfig|Prune|null
     {
         return match ($this) {
             self::KeepForever => null,
-            self::PruneOlderThan30Days => new PruneConfig(before: Carbon::now()->subDays(30)),
-            self::KeepLast5 => new PruneConfig(keep: 5, varyOnData: false),
-            self::KeepLast3VaryOnData => new PruneConfig(keep: 3, varyOnData: true),
+            self::PruneOlderThan30Days => Prune::before(Carbon::now()->subDays(30)),
+            self::KeepLast5 => Prune::keep(5)->dontVaryOnData(),
+            self::KeepLast3VaryOnData => Prune::keep(3)->varyOnData(),
             self::NoPruneConfig => null,
         };
     }